Get_Next_Line

+++ 2-Get_Next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: ebodart <ebodart@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 10:32:35 by ebodart           #+#    #+#             */
-/*   Updated: 2021/02/25 12:16:56 by ebodart          ###   ########.fr       */
+/*   Updated: 2021/02/25 12:44:25 by ebodart          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@
 # include <unistd.h>
 
 int		read_line(char *keep, int fd);
-int		fill_line(char **line, int size, char *keep, int ret_read);
+int		fill_line(char **line,long int size, char *keep, int ret_read);
 int		size_line(char **line, char *keep, int ret_read);
 char	*join_keep_line(char *tmp, char *line, int i);
 int		get_next_line(int fd, char **line);
 
-void	init_keep(char *keep, int i);
+void	init_keep(char *keep, long int i);
 size_t	ft_strlen(const char *s);
 void	*ft_calloc(size_t nmemb, size_t size);
